Refactor, the pakages DataAcces, BusinessLogic and Presentation was created
